Update pbr: warn_dhcp_force
Display a warning if user has dhcp.force = 0 (disabled). If disabled a request is put out to identify other DHCP-servers on the network. This takes up time (about 3s ¬ 6s depending on start/stop/restart) and is uncommon for most users.

I could not find a good place to put it because the code is still unfamiliar., so its added to restart to give an idea how it would function.