pbr: is_supported_interface update
Present code logic will process an interface if it is listed as supported even if it is disabled.

To me it makes more sense that if an interface is disabled it will not be processed at all.

This patch should make that so.

If you agree then please review and implement, otherwise just discard and close

Signed-off-by: Erik Conijn egc112@msn.com