pbr: netifd set ipv4 and ipv6 default route for pbr netifd tables
Still struggling with netifd.
If you have a VPN as default route and you stop the VPN the default route will still try to go through the VPN as that is not reversed.
Furthermore I see my endpoint for the VPN routed via another VPN whhich is not working.

For now it looks like with netifd it only works reliably if you keep the WAN as default.
That is now possible with this patch as there will be a default route made in the pbr netifd tables
But still this behaviour is different from without netifd.

Use this patch as you seem fit, change or move it etc. if you want

Tested on R7800  23.05.5 NSS build

Signed-off-by: Erik Conijn egc112@msn.com