pbr: ip rule delete and sport patch
This patch should solve a problem where the ip rules, in particularly the ipv6 rules are not properly deleted
One problem is that `interface_routing 'destroy'` does not send the priority number so that the rules are not deleted at all.

For non netifd tables the IPv4 rules are flushed later on but not for IPv6.

By fixing this we should be able to use the same priority number for ipv6 instead of using `$((priority-1))`

Second problem was that wg servers sport where not set and destroyed properly for ipv6 this should also be fixed.

Tested on my R7800 build 23.05.5 nss build

Please have a look and consider implementing

Thanks for all your great work

Signed-off-by: Erik Conijn egc112@msn.com