pbr: is_domain, allow hyphen
The current implementation does not allow a hyphen.
This patch allows one or more hyphens (in addition to alphanumeric characters) for second and subsequent characters.

A hyphen is allowed but cannot be the first or the last character and although my patch does allow a hyphen as last character, at least it allows a hyphen, so maybe not optimal but at least better.
The complete validation is more complex but as this is only used to make a distinction between ip address and domain name, it could be sufficient.

REF: https://cs.uwaterloo.ca/twiki/view/CF/HostNamingRules

Please have a look and consider implementing

Signed-off-by: Erik Conijn egc112@msn.com