pbr: enhance status
`service pbr status` only displays IPv4 and not IPv6 tables and rules.
This seems to be caused by the fact that `$ipv6_enabled` is not available.
I devised a workaround getting the variable from the config with `uci_get`
Furthermore I changed the code to show the IPv6 rule(s) and added not only the table number but also the table name and added some line breaks for better readability.

Please have a look and consider implementing, feel free to change/improve this pull request 😊

Tested on R7800 running 23.05.5

Signed-off-by: Erik Conijn egc112@msn.com