pbr: wg-servers ip-rules not reloaded on reloading interfaces
Maintainer: @stangri

Tested: on Dynalink DL-WRX36 running main build

Description:
Problem reported by: https://forum.openwrt.org/t/policy-based-routing-pbr-package-discussion/140639/2057?u=egc

I also have it sometimes after a reboot when interfaces are still restarting, PBR will reload the interface but not the wg-server ip rules.

This problem can be triggered by doing `service network restart` this will not reinstate the wg-server ip rules.

When doing a `service pbr restart` it works again and the wg-server ip rules are in place.

A possible solution is to reinstate the wg-server iprules not only on boot/restart but also on reloading of interfaces.

After some research the "action" is `reload_interface` so adding that to
https://github.com/stangri/pbr/blob/43bdaa096d04883a08d6d1b20cd55248805ed2aa/files/etc/init.d/pbr#L1774 might solve it

Furthermore maybe `reload_interface` should also be added to the tor interface in line?:
https://github.com/stangri/pbr/blob/43bdaa096d04883a08d6d1b20cd55248805ed2aa/files/etc/init.d/pbr#L1755

Signed-off-by: Erik Conijn egc112@msn.com