pbr: in r18 is_domain does not differentiate between ipv4 and domain name
Maintainer: @stangri

Tested: on Dynalink DL-WRX36 running main build

Description:
In r18 the is_domain() function does not diffentiate between ipv4 address and a domain name
This will cause trouble when domain names and IPv4 addresses are used.

As a possible solution I have added the check for `is_ipv4`.
This might not be the best possible solution so feel free to do it otherwise.

But it is a problem which needs to be solved and with this pull request it at least comes on the TODO list :)

Signed-off-by: Erik Conijn egc112@msn.com